+++ src/telink/hal/zigbee_zcl.c
@@ -50,7 +50,7 @@ static cluster_registerFunc_t get_register_func_by_cluster_id(u16 cluster_id) {
   return NULL;
 }
 
-static status_t cmd_callback(u8 endpoint, u8 clusterId, u8 cmdId,
+static status_t cmd_callback(u8 endpoint, u16 clusterId, u8 cmdId,
                              void *cmdPayload) {
   hal_zigbee_cluster *cluster = hal_zigbee_find_cluster(
       hal_endpoints, hal_endpoints_cnt, endpoint, clusterId);
@@ -166,7 +166,7 @@ void telink_zigbee_hal_zcl_init(hal_zigbee_endpoint *endpoints,
   }
 }
 
-void hal_zigbee_notify_attribute_changed(uint8_t endpoint, uint8_t cluster_id,
+void hal_zigbee_notify_attribute_changed(uint8_t endpoint, uint16_t cluster_id,
                                          uint16_t attribute_id) {
   report_handler(); // Trigger reporting if needed
 }
